backend/requirements.txt
Flask==3.0.2
Werkzeug==3.0.1
flask-cors==4.0.0
python-dotenv==1.0.1